ml_cicd/custom_train_validation/requirements.txt
# requirements.txt
scikit-learn==1.1.3
joblib==1.2.0
google-cloud-storage==2.7.0
numpy==1.24.1
pandas==1.5.2
google-cloud-bigquery==3.4.1
db-dtypes==1.0.5  # Good to have for BigQuery/Pandas integration

gcsfs==2022.11.0
